refactor(DiscussionScreenTest): Separate unit tests from integration tests into two files
